Telemetry/RF Telems/onboard/requirements.txt
Polygon2
TTFQuery
VPython
crcmod
fonttools
matplotlib
mock
nose
numpy
pyglet
pyparsing
pyserial
python-dateutil
pytz
pyzmq
selenium
six
vboxapi
wsgiref
wxPython
wxPython-common
